Started to move builder help functions from apio_env.py to plugin_util.py. Also, removed the apio_env.py logging methods msg, warning, error, etc. Using secho directly is easy enough.
