+++ apio/scons/plugin_gowin.py
@@ -58,7 +58,7 @@ class PluginGowin(PluginBase):
             ),
             suffix=".json",
             src_suffix=".v",
-            source_scanner=apio_env.verilog_src_scanner(),
+            source_scanner=self.verilog_src_scanner,
         )
 
     # @overrides
@@ -141,7 +141,7 @@ class PluginGowin(PluginBase):
             generator=action_generator,
             suffix=".out",
             src_suffix=".v",
-            source_scanner=apio_env.verilog_src_scanner(),
+            source_scanner=self.verilog_src_scanner,
         )
 
     # @overrides
@@ -172,5 +172,5 @@ class PluginGowin(PluginBase):
                 lib_dirs=[self.yosys_lib_dir],
             ),
             src_suffix=".v",
-            source_scanner=apio_env.verilog_src_scanner(),
+            source_scanner=self.verilog_src_scanner,
         )
