+++ apio/scons/plugin_ice40.py
@@ -58,7 +58,7 @@ class PluginIce40(PluginBase):
             ),
             suffix=".json",
             src_suffix=".v",
-            source_scanner=apio_env.verilog_src_scanner(),
+            source_scanner=self.verilog_src_scanner,
         )
 
     # @overrides
@@ -138,7 +138,7 @@ class PluginIce40(PluginBase):
             generator=action_generator,
             suffix=".out",
             src_suffix=".v",
-            source_scanner=apio_env.verilog_src_scanner(),
+            source_scanner=self.verilog_src_scanner,
         )
 
     # @overrides
@@ -170,5 +170,5 @@ class PluginIce40(PluginBase):
                 lib_files=[self.yosys_lib_file],
             ),
             src_suffix=".v",
-            source_scanner=apio_env.verilog_src_scanner(),
+            source_scanner=self.verilog_src_scanner,
         )
